Middleware/models/stage2/requirements.txt
# Core Dependencies
torch>=2.0.0,<3.0.0
torchvision>=0.15.0
detectron2>=0.6
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
openpyxl>=3.6.0
PyYAML>=5.4.0
albumentations>=1.1.0
Pillow>=8.3.0
matplotlib>=3.4.0
